+++ models/comment.go
@@ -1,0 +1,17 @@
+package models
+
+import "time"
+
+type Comment struct {
+	CommentID int       `json:"comment_id"`
+	ArticleID int       `json:"article_id"`
+	UserID    int       `json:"user_id"`
+	UserName  string    `json:"user_name"`
+	Message   string    `json:"message"`
+	CreatedAt time.Time `json:"created_at"`
+	UpdatedAt time.Time `json:"updated_at"`
+}
+
+type UpdateComment struct {
+	Message string `json:"message"`
+}
